LPS-93514 Move LayoutPortletKeys back to layout-test, cause it is only used in layout-test, no need to put it in layout-test-util. Rename the package to avoid name conflicts.
