+++ modules/apps/layout/layout-test/src/testIntegration/java/com/liferay/layout/test/LayoutCopyHelperTest.java
@@ -20,7 +20,7 @@ import com.liferay.fragment.service.FragmentEntryLinkLocalService;
 import com.liferay.layout.page.template.service.LayoutPageTemplateStructureLocalService;
 import com.liferay.layout.page.template.util.LayoutPageTemplateStructureHelperUtil;
 import com.liferay.layout.test.util.LayoutTestUtil;
-import com.liferay.layout.test.util.constants.LayoutPortletKeys;
+import com.liferay.layout.test.constants.LayoutPortletKeys;
 import com.liferay.layout.util.LayoutCopyHelper;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.json.JSONObject;
