+++ modules/apps/layout/layout-test/src/testIntegration/java/com/liferay/layout/test/constants/LayoutPortletKeys.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.layout.test.util.constants;
+package com.liferay.layout.test.constants;
 
 /**
  * @author Eudaldo Alonso
